auth_api/python/requirements.txt
Flask
PyJWT==1.4.2
cryptography==3.1.1
mysql.connector